Completion of assembly strategy factory design
* This change was made to produce flexible and modular machinery to
allow the DocumentGenerator instance handling a document resource
request to be able to intelligently choose the correct assembly
strategy not just based on the assembly strategy chosen, e.g., verse,
but also based on which resources were actually requested and found.
The whole idea behind this architecture is to make assembly strategies
modular and mostly free of conditional logic that would otherwise be
present to handle the various combinations of resources that could
occur.

For example, a user could request English USFM and Translations
Notes (TN) for Luke, Brazilian Portuguese for same, and
Kishwali (Swahili) of USFM and Translation Notes for Colossians. It
turns out that the English USFM and Translation Notes are indeed
available and thus found, and same for Brazilian Portuguese, but
Kishwali USFM is found while Translation Notes are not since they do
not yet exist. Notice that for this particular document request, USFM
and Translation Notes were requested, but other combinations of
resources were possible that were not requested, e.g., USFM only, or,
TN only, or USFM, TN, plus TQ and TA, per language and per book etc..
Imagine the conditionals otherwise required to handle such
combinations when it comes time to assemble all these resources into a
document.

If the assembly strategy was coded as a big ball of mud with
conditionals everywhere to handle such contingencies then it would be
an ugly, unmaintainable, and, importantly, un-extendable mess. But
with this design the system makes the decision on what assembly
strategy to run by simply sending the (Optional) resource instances,
if they exist or None otherwise, of each resource to the assembly
strategy factory and the factory does the (relatively simple) work of
dispatching to the correct assembly strategy (a higher order function)
in a lookup table based on what resources "exist" which in our terms
means it was both requested, available, and physically found in the
cloud. Thus, each combination of resources that we support runs an
assembly strategy that is unique to that combination. This makes
assembly strategies as lean as possible and specialized leading to
much better cognitive understanding of what is going on and why. The
idea for this solution comes from type level programming ideas in
Haskell (if you are interested), but really is just a factory
pattern in OO land.

Note that at this time, with this commit, we support completely the
USFM and TN combination of resources. Coming are strategies that will
handle the other combinations we wish first to support, e.g., USFM
only, TN only. Later, when we support Translation Words (TW),
Translation Questions (TQ), and Translation Answers (TA), and others,
their assembly strategies will be added also.

Book and chapter interleaving strategies have not yet been identified
as a requirement, but are certainly possible. So they have been
retired for now, but can now be easily resurrected.