Log resource types that were requested but don't exist
Add code that can provide the resource types (and maybe later the
resource codes) that are available for a particular lanuage code. Use
that code, for now, to make it possible to log resource requests that
don't exist. This code is experimental, but is aimed at tracking, for
now, what options should not be presented to users in the BIEL user
interface. Further development can proceed to add to the resource
types, the available resource codes for a language and could possibly
then be used by BIEL to construct its UI menus such that users are
never presented with options the system cannot fulfill. I.e., better
UX.