WIP: Delay initialization of verses HTML until assembly
Delay initialization of non-USFM resources' verses HTML until assembly
strategy time so that translation word file paths can be passed to
each initialize_verses_html call for each non-USFM resource. This
turns out not to work because the TranslationWordLinkExtension does
not accept the frozenset of Markdown file paths for some reason even
though it is hashable. I have not been able to uncover why. A
different approach will appear in the next related commit and this
commit is documentation of this attempt.