Update Makefile
* Add 'all-plus-linting' target.
* Update 'all' target to not include mypy and linting.

Keep 'make all' relegated to building and testing only as I've
experienced some issues running mypy and the linters in a git worktree,
i.e., in a clean checkout of the project for reasons I haven't yet taken
time to determine. So, mypy and linting works well on my dev machine but
could cause a failure on other machines.