Don't require that at least one resource was found
It is possible to have not found any resources due to a malformed
document request, e.g., asking for a resource that doesn't exist as
asked (in translations.json). Thus we can't assert that it always
exists as a non-empty list as an input param. The system already
handles missing resources fine, so there is no need for this check.
Any references to tickets, articles, etc?