Put working_dir and output_dir back to type str
There was an issue with pydantic.DirectoryPath type wherein Python
complained that attribute _flavour was not implemented. This happens
because DirectoryPath is a subclass of Path which itself doesn't
implement the _flavours. It seems that it may be a bug only when
instrumenting the system with typeguard package. There are some links
online to others running into this and what I consider unsatisfactory
resolutions/hacks to get around it. So, back to str we go. We can
always assert os.path.exists on working_dir and output_dir if we wish
to get some of the same behavior of pydantic.DirectoryPath.