Add gunicorn config, update Docker config, update logging config
This commit adds a Gunicorn management process which manages Uvicorn
worker processes and handles load balancing between them. Gunicorn is
configured to choose the optimal number of workers based on the
specifics of the server on which it is run - number of CPUs and cores
per CPU. I.e., Gunicorn configures itself dynamically, but its
configuration also allows configuration values to be overridden using
env vars.

Additionally this commit modernizes the Python configuration in that
it utilizes the exact same virtual environment as used in development.
Many think virtual envs are not useful or required in Docker
containers, but actually there are several benefits and those benefits
are leveraged here.

This commit also updates the logging config has been updated to remove
the file handler which was recently added to provide a
RotatingFileHandler for logging. The reason is that in the deployment
environment for this project, logging only to standard out suffices so
there is no need to take up disk space with rotated logs.

Lastly, many Makefile targets have been updated.