Make it possible to update a resource asset git repo if caching period expires
Every other type of resource asset could have its cache invalidated
but git repos. Until now, they would be cloned but subsequently never
updated. What we do now when the caching period has expired for a
resource asset git repo is to delete the directory of the repo and
then clone it again.