Update Makefile targets
Separate the installation of mypy type stubs into its own Makefile
target and make it a dependency of mypy Makefile target.
Modularization.

Update two Makefile targets

A prior commit had used local-prepare-for-tests-without-cleaning to
allow for testing the caching mechanism because that target does not
remove generated PDFs prior to a test run. Now that that caching is
verified to be working, go back to version local-prepare-for-tests
which does remove generated PDFs from previous test runs prior to
running the test.