Add clean-mypyc-artifacts Makefile target
mypyc creates C binaries. When making
changes and verifying them in a quick edit-test-run development
workflow having to recompile using mypyc in between test runs slows
down the flow of work. This Makefile target can be used to quickly
remove the generated binaries so that we are back to running
interpreted Python temporarily. Then when we have our tests passing
and changes we can go back to compiling source with mypyc via make
mypyc.