Show the max number of chapters and verses available
Prior to this commit you could get an undesirable state of affairs in
book then language assembly wherein if the first usfm language/book
resource was missing a verse, then other usfm language/book resources
in that same document request would skip that verse because he
chapter and verse loops were arbitrarily based off the first usfm
resource. Now the usfm resource with the most chapters for a
particular book and subsequently the usfm resource with the most
verses within each respective chapter are used as the outer loops
pumps for chapters and verses repsectively. This way the max available
chapters and verses are displayed in the resulting document.