Stricter adherence to functional style
Don't use "globals" even if they are just a module reference when it
can be passed as a defaulted function argument instead. Pass in what
we need as parameters to keep things more pure. I say more because we
aren't yet to a place with the code that doesn't have side-effects,
that is coming though.