Partial verses bug fix (#9)
* Fix for partial verse display

Partial verse display was reported as a bug.

* Development-side ergonomics

Various tweaks to make working on the development side more efficient.

* Updates to pip packages

Automatically updated pip packages as a result of build process.

* Add some type casting to make mypy happy

mypy and thus mypyc build phases were failing due to mypy being unsure
of the types.

* Fix additional issue that bug fix introduced

Due to complexities of parsing HTML that is generated from USFM and thus
mirrors the complexity of the USFM spec it can be the case for some
language/book combinations, e.g., Jonah 2:2, that verses adjacent to the
current verse associated with a verse reference label may appear in
appended position to the current verse. Handle this case without
regressing others, without interfering with the analagous case wherein
unwanted verse span tags may appear in prepended position to the current
verse (something already fixed prior to this commit).