Refactor remove_section_preprocessor module and update existing test
Refactor remove_section_preprocessor to follow the pattern that the
other markdown preprocessor plugins follow so that the parameters the
plugin needs are passed in at construction time for the sake of
referential transparency.

Existing test was improved to ensure that markdown sections which are
not to be removed and occur after sections that are to be removed
remain intact after processing by this preprocessor.

Regex approach used in remove_md_section was abandoned in favor of
legacy algo since regex complexity required to yield desired result is
arguably more complex to grok for newcomers.