Choose better Makefile target dependency
We really need requirements.txt and requirements-prod.txt to be
updated before building the (local, non-production) Docker image as it
depends on this. The existing Makefile target dependency prior to this
commit already did that. However, let's also require that the packages
specified in requirements.txt and requirements-prod.txt get installed
locally, i.e., outside Docker, correctly before attempting a build as
an integral part of the build as a sort of precheck since the Docker
image build will do the same itself, but if it fails as a precheck we
will have saved some time and fail early.

It should be noted if it wasn't already obvious that this has no
impact on the production build.