Remove or modify certain integration tests for sl, sr layout
The select layout algorithm is more robust now and certain test cases
no longer make sense. Also we reorganize a few tests in terms of the
order of their resource requests so that the order is usfm, tn, tq,
tw, etc.. Such order is not required, just a develoioper side
organization.