Add ePub and Docx support
Pandoc handles conversion from HTML. This commit adds necessary logic
and does a bit of required refactoring at the same time. Handling of
attachments becomes more generic so that it is not tied to PDF.
Frontend provides checkboxes for PDF, ePub, and Docx. UI is still v1,
update comes in a later PR than this.