WIP Use daisyui rather than smui
Tailwind, postcss, and daisyui (which is built on them) is a far more
flexible wrt styling for out path to the v2 UI. This commit doesn't
get the styling all the way there, but it gets closer. The new v2 ui
is still in the works, this are just building blocks.