Languages, resource codes, resource types pages working
Storage of state now works in all three pages including with return to
page after navigating away from it in-site.

Moved away from Svelte's #await blocks in favor of putting more
Typescript logic inside the script tag for three reasons:

    1. Currently Svelte does not support Typescript preprocessing in the
    template itself, only inside the script tag and we are using
    --strict Typescript so this is problematic.
    2. It is better to keep most logic out of view oriented code, i.e.,
    out of templates.
    3. I do not like the common pattern in Svelte of side-effecting
    component-global variables inside an async function. This breaks
    referential transparency which is a bad thing from a design
    perspective. I instead moved to a Promise resolution pattern in the
    script tag and then use #if statements in the template. Also moved
    to smarter usage of reactive variable initialization to keep
    template code mostly free of logic comparatively to before.

Still upcoming is the ability to have intra-store reactivity so that
when the user returns to a previous page to update their choices in a
way that invalidates previous choices, the UI will identify that and
notify the user with appropriate UX cues per our design doc.