Reverse sense of test
Language kbt, Abadi has an invalid URL in translations.json so we
expect tests which used to succeed to now fail.

Due to this, we are also currently filtering that language out of
available choices. When it is fixed we can make it available again.
There are likely many languages which we will find that have invalid
URLs in translations.json.