Experimental code that I want to save
The idea here was to get the remaining data at the same time as when
the user chooses their two languages using the time the user expends
interacting with the UI between hitting submit and navigating to the
next step. I may return to this depending on yet to be determined
circumstances, but based on testing I will likely not use this code
later. Still, I will keep it for now as it has many good ideas in it
that are hard to elaborate here.