Use text-based indeterminate progress indicator for now
To use the graphical progress indicator there is work to be done, if
we decide it is worth it, to emit backend progress events and then
listen for them on a websocket to update the progress bar. It may
still happen, but just hasn't yet as it is relatively low priority
compared to other things in this update.