Development-side makefile target for testing a specific case
kbt was bombing and it turned out that rather than a bug, it has an
invalid URL in translations.json. Subsequently, I have found at least
one other with the same problem and I suspect there are many more. The
code handles this and reports error to user per visual design spec,
but (as mentioned in another earlier commit) these types of issues
should be (automatically) reported to upstream rather than allowing
them to persist and downgrade UX.