For now, don't try to determine max verses in a two language setting
Languages don't necessarily have the same number of chapter and verse
content for the same books across languages. I've made consistent
efforts to handle such situations by always choosing as a chapter and
verse "pump" (for iteration) the book/language that has the max number
of chapters and then (in an inner loop) max number of verses for same
book. This works well, but I want to handle exceptions and the way
logic is layed out with respect to this a little differently later.
This commit ensures using the max chapter pump, but then uses that
book's verse "pump" for iteration rather than attempting to also get
the book within that chapter that has the most verses as the verse
"pump" because it causes no KeyError exceptions due to chapter_num not
being a key into a book that may not have that chapter. Mostly just a
note to myself.