Add/improve source comments
Indicate which functions are available via API but not through UI to
avoid any confusion for those reading the source code. Head of IT
wanted number of choices slimmed down and discussion led to these
functions not being used through UI. Vestigial at this point, but
still valuable.

Change FIXME to different status.