Docx support and v1 UI candidate featuring basic and full uis (#91)
* Add initial proper docx support

Docx is the new preferred document output format as it is more easily
editable for translators.

This commit adds native docx support and an associated path for it in
the UI and backend as well as a test suite for it.

Supporting Docx meant adding a new code path that doesn't convert from
a finished HTML document to Docx in a wholesale fashion (in comparison
with conversion to PDF or ePub). Instead, the Docx is built up from
HTML snippets intermingled with required low-level Docx manipulations
to control one or two column layout and ultimately Docx template use
too. In turn this means that certain UI changes were necessary. The UI
changes include making the user choose either PDF, ePub, or Docx
rather than being able to choose all of them at once if desired. This
has the added benefit that a user can no longer choose to create three
different output formats in one go and thus the time for the request
to be fulfilled will be greatly reduced in such cases which will also
make the user feel the system is snappier (objectively, but more
importantly, subjectively).

Another UI change not directly related to Docx support is that we have
removed the drop down control to select chunking size. Instead, the UI
always sends chapter chunk size now and never verse chunking. Verse
chunking was decided to be of no value for translation purposes at
this time and verse interleaving is very costly in terms of
performance which makes for a bad UX. It can be resuscitated from git
history later if we want it back in the future though it is hard to
imagine now that happening.

Note too that Docx support does not include one of the layout options
available to HTML, PDF, and ePub: two column side by side languages.
This was never vetted as a needed requirement and has not been used or
requested for some time. It still exists as an option in HTML, PDF,
and ePub output formats as sort of a showcase for what can be done.
Also this layout would be much more difficult to implement in Docx.

This commit does not yet accommodate fonts for certain languages in
Docx format. To do so in Word will have to be the subject of a future
PR and appears to be of some complication. More research is needed.

There are also a few miscellaneous commits related to better naming,
a fix to configuration in docker-compose-override.yml, changing chunking
of several tests from the no longer used verse chunk size to chapter
chunk size instead, and remove commented out link in a Jinja HTML
template.

In a coming PR we will likely also want to ensure sources present no
linking in the generated doc. For now, for some sources, there are
some dead links in PDF, ePub, and Docx output formats. There is a
great deal of heterogeneity in links in the source. We successfully
handle tens of types, but there are apparently yet more that exist and
sometimes they are just malformed in the sources.

* Update non-docx assembly strategies interleaving order

Update the interleaving strategies for HTML, PDF, and ePub to follow
the order of interleaving that is implemented in the docx assembly
strategies which were designed to adhere to the product owner's latest
requirements updates as exemplified in a Word file that acts as
template. Specifically, book intro materials (TN and BC - BC wasn't
presented in the example, but intuition says it could follow TN book
intro) followed by USFM and footnotes, followed by chapter intro,
followed by verse-level translation notes, followed by other resource
assets (TQ, secondary USFM), and finally TW at language level. The old
order was book intros, chapter intro, USFM, verse-level TN notes, then
other remaining resource assets (with some variation).

* Update Makefile

* v1 UI candidate and full ui version both accessible and linked

- [x] basic version, i.e., v1, is available at #/v1/.

- [x] full version is available at #/

- [x] You can switch between them using a link in the About page of either

- [x] The Mast (the top area of the app where the logo and tab menu
  items live can be shown or hidden based on a bujild-time configuration
  setting in source at DOC/frontend/.env: VITE_SHOW_TOP_MATTER (which
  controls this for the full version) and
  VITE_SHOW_TOP_MATTER_IN_V1 (which controls this for the basic, v1,
  version). The usefulness of this comes into play if you want to, say,
  display the app without the Mast so as to integrate into an iframe
  within BIEL. If the app lives at its own link (with respect to
  navigation) then likely you would want the Mast kept in place.

- [x] Similarly, in the basic ui version, v1 candidate, there is a
  flag in the same .env file as above for showing or not showing
  resource types. We choose the approved USFM and TN types that are
  also listed in translations.json for the user. As such there is not
  a functional reason to show the resource types to the user. However, it
  likely a less surprising UX for the user if they are at
  least shown what resource types the system chose for them. The
  product owner will decide which option is best.

- [x] We didn’t yet actually have anything useful in the
  hamburger menu at the upper right in the Mast.  Thus, for now, I
  have disabled the ‘Hamburger ‘ menu in the top right of the Mast in
  both the basic and full UIs.

- [x] For now, the basic version, i.e., v1, defaults to PDF output.
  Why? Because many gateway languages use special fonts. As you
  likely recall from recent emails the handling of such fonts in
  Word document generation, especially for multi-language
  documents, is yet unsolved. I do have a lead on a couple
  approaches that may prove useful, but there is more work to be
  done there which will be the subject of a future PR or update to
  this PR if it has not yet been merged when I return to it.
  In the meantime, if you want to see the docx output for a gateway
  language simply switch to the full ui version (using the
  link in the About page or by manipulating the browser location URL
  to #/v1/) and request to generate Docx for said gateway language and
  view the result. For many gateway languages the result will be just
  fine, but for many which have special fonts you’ll get a document
  with fine structure (unless the underlying resource source assets
  are malformed or incomplete - this is true of many languages)
  but filled with unicode-font-missing boxes. Some backstory: Our PDF
  output comes from HTML that is rendered headless by webkit and the
  browser handles the fonts transparently (assuming they are installed
  into the environment the browser is running in), whereas with Docx
  generation things are more complicated.

    - [ ] What is likely going to be needed is a map of language to
      font needed for each language as well as the language’s
      identifying code (from
      Word’s perspective -probably an ANSI code or omething - see
      https://stackoverflow.com/questions/36967416/how-can-i-set-the-language-in-text-with-python-docx
      for example). For v1, we’ll only need that just for the gateway
      languages, but later for all languages. Hopefully there is a
      different less tedious way to solve this, but the link provided
      is the closest I have come so far as an approach that might work.
      With more time for researching it might turn out to be less
      onerous. Let's hope so.

- [ ] Other more minor items like font darkness/lightness, etc on
  various controls will get my attention before the deadline for
  launch.