Update3242023 (#94)
* Move pytest config to pyproject.toml

Modern config style for pytest which moves config from pytest.ini to a
section in pyproject.toml. Seems to work more reliably too.

* Fix non-balanced columnar document layout and improve PDF layout

This is achieved through switching PDF engines to WeasyPrint from
wkhtmltopdf. WeasyPrint understands CSS column-count directive whereas
wkhtmltopdf did not. WeasyPrint finally acquired header and footer
support so this made the switch finally possible.

* Fix bug wherein docx stayed set in UI on subsequent request

* Set default for assembly strategy rather than give choice

Previous to this commit the UI allowed a user to choose to 'separate'
or 'mix' content. Now content is set to always separate.

This means you can still request multiple (a max of two) languages in
one document, but you'll see all the content for one language and then
all the content for the other language rather than mixing one
chapter (scripture and helps) of one language and then a chapter of
another language.

The 'mix' strategy could be brought back later, but mostly existed to
allow the user to choose the scripture left, helps right layout
strategy which has since been put aside as a non-requirement.

The net effect of this commit then is to also simplify the UI yet
again for the full version of the UI so that the user has one less
choice to make and yet the documents produced still satisfy all
requirements we have heard expressed thus far.

* Update Docx support

- Reliably provide title1 and title2 for the Word template first page
- Fix/improve tests for Docx
- Update docx template headers to pick up appropriate headers (this
  still needs one more commit after this to complete fully - some of
  the content still needs some heading level adjustments)
- Update Results download page to provide some guidance to user
  regarding fonts: where to acquire them, etc.
- Etc.

* Remove obselete commented out code in email template

* Fix developer-side docker compose file for Docx test invocation

* Fix test to produce valid document request

* Updates to Makefile for developer-side

* Comment out UI tests that no longer apply

Scripture left, helps right is not an available layout strategy in
this version of DOC.