Update mar 30 2023 (#95)
* Handle Bahasa Indonesian language

Language code ID USFM is not included in translations.json due to
licensing and other issues. It is still a requirement to have ID
supported. This commit provides a hardcoded (yet otherwise well
designed) workaround for handling ID.

* Uncomment scripture left, helps right layout strategies

Today we learned that this strategy may officially become an actual
requirement. Uncommenting the code for these strategies short of the
UI means that at least we can test them through the API. They still
work but have undergone a (hopefully) minimal amount of visual
breakage detected during manual testing that needs to be cleaned up
before appearing as an option in the UI again. Their (non-visual)
tests all still pass.

* Add underlining to a couple links in About pages

Add underlining to a couple links in About pages for basic and full
versions of the app so that visually the links are more obvious for a
better UX.

* By request, reduce the vertical space after a heading

In translation notes sections, the bolded headings, h5 headings, were
thought to have excessive vertical white space below them. This commit
reduces that vertical white space to a hopefully acceptable distance.

* Fix for bug laying out TW content in Docx format

A previous commit forgot to make these changes despite all tests
passing (which means no adequate test was in place to cover this
case). Manual testing of a particular case discovered this.

* Don't enable scripture left, helps right layout at this time

With scripture left, helps right being a non-requirement for this
release we've disabled it. We learned today that it might likely be a
future requirement. Because a visual defect in the layout was
discovered recently we'll need to do some work on it and enable it
again for a future release if it does indeed become a required layout
option.