Fix docx rendering issue
By experimentation is what found that our 3rd party HTML to Docx
rendering library does not like having spans not enclosed in a block
level HTML element. For many languages (possibly thousands), typically
non GL languages, it is not uncommon for the USFM to have a USFM
chapter marker followed, even on the same line (a violation of the
USFM spec), by USFM verse markers.

When this construct gets parsed to HTML by our usfm_tools
module (which we maintain), it means that the resulting verse spans
are not enclosed by any block level element other than the body
element.

For PDF and HTML this is handled fine as is, but the libraries we use
for Docx rendering take this to mean that the verse spans are a
continuation of the chapter marker resulting in a visual defect in
which all the verses appear in the styling of the chapter headline
rather than as normal body text/style.

A fix that was found in practice is to insert a USFM paragraph marker
right after the chapter marker in such cases which, when parsed,
results in all verse spans being enclosed in a block level HTML
paragraph element. This fixes the visual defect.

Tests were carried out manually on Aushi and Abuhainu languages, which
had this defect, and now they are rendering properly. All other tests
passed. It is likely that this fix solves the problem not just for
Aushi and Abuhainu, but potentially thousands of other languages yet
untested.