Fix issue with mr language content handling
Turns out that regex search, which is part of our limit tw words
feature, on mr language content (specifically tw resource content) was
failing because of the regex expression was not escaped and that
expression is actual mr content (specifically the localized tw word
itself) and some mr tw words have a parenthesis in them with no
matching closing paren or vice versa and so since open and closed
parens are interpreted in the regex context as control characters it
would fail saying it was an invalid regex expression. Escaping the
regex expression using re.escape fixed this issue for mr language, but
also for any other language that has tw words containing regex control
characters.

Update test suite to no longer skip tests that were failing due to
this defect.