Fix bug in asset file layout discovery
Initially this presented as a bug wherein language Aushi and mat, mrk,
luk, acts chosen with only mat and luk showing in final result
document. Further investigation led to finding a naive assumption in
the logic of the code about the location of acquired assets for the
selected resources. This fix removes this naive assumption and
replaces it with code that makes no assumptions that are not
supportable by the reality of resource asset file layout and naming.