Nicer language name presentations
In translation.json a language has a localized name and its English
name. Unfortunately, many languages have included the English
name (usually in parens) in the slot designated for localized name
only. Further we display the English name next to the localized name
in the UI by requirement and so for those languages with this data
problem they end up having a name presentation like 'Foo (Bar) (Bar)'.
So this commit adds some logic to detect if the English name has been
erroneously included in the localized name and if it has it does not
add the English name in parens (since it is already there), otherwise
it does add the English name.