Make file name more semantic wrt TW
Previously, document file name ended in ltwt or ltwf depending on if
TW words were limited to those that occur in book (by choice of user)
or if TW words were not limited or if TW was not present at all
respectively. A better semantic is to only add ltwt or ltwf to the end
of the document file name if TW was actually requested in the
document (for the ltwf case). Since file names are also lookup keys
in our file cache this seemed a worthy upgrade, but more important is
the semantic import of the file name communicating what is in the
document request.