Fix email issue properly with pydantic
A few tweaks were necessary to make type checked env vars work with
pydantic library again:

- Add email/SMTP env vars to api stanza in docker-compose.yml
- Add email/SMTP env vars to test-runner stanza in docker-compose.api-test.yml
- Etc, see diff for details

Why were these changes necessary? We instantiate pydantic's Settings
class to provide configuration. The Settings instance is used in api,
worker, and test-runner services and not just for SMTP secrets.
Because the SMTP secrets/env vars were missing in api service and
test-runner services the instantiation would fail since we were
requiring them as attributes on Settings class, i.e., in order to use
pydantic to supply the env var values we needed these additions. If we
didn't want pydantic's features, mainly type checking, we wouldn't
need to supply SMTP env var secrets anywhere but the worker service
stanza in docker-compose.yml as long as we used os.environ["name of
var"] to access them, but this is not desired.