Downgrade back to previous version of playwright for Docker
Seeing if this will solve an error on github action complaining of:

"External plugin loader failed. This most likely means that the plugin
loader helper binary was not found or could not be run. You might need
to set the GST_PLUGIN_SCANNER environment variable if your setup is
unusual."